display username on load
