+++ src/index.html
@@ -23,7 +23,7 @@
     <div class="logo-wrapper">
       <div class="admin-wrapper">
         <button id="admin" class="admin-button">A</button>
-        <p>Admin Username</p>
+        <p id="username">Admin Username</p>
       </div>
       <p class="logo" tabindex="0">TravelTracker</p>
     </div>
